Changed signatures of bbox and draw to reduce use of floats

+++ src/org/meta_environment/rascal/library/experiments/VL/Grid.java
@@ -22,7 +22,7 @@ public class Grid extends Compose {
 	}
 	
 	@Override
-	void bbox(float left, float top){
+	void bbox(int left, int top){
 		this.left = left;
 		this.top = top;
 		width = getWidthProperty();
